refactor(agent): decouple agent and ovsdb monitor
