+++ pkg/monitor/agent.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 	"net"
 	"os"
-	"reflect"
 	"sync"
 	"time"
 
@@ -35,7 +34,6 @@ import (
 	"k8s.io/klog"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/everoute/everoute/pkg/agent/datapath"
 	agentv1alpha1 "github.com/everoute/everoute/pkg/apis/agent/v1alpha1"
 	"github.com/everoute/everoute/pkg/types"
 	"github.com/everoute/everoute/pkg/utils"
@@ -46,119 +44,49 @@ const (
 	AgentInfoSyncInterval = 5
 )
 
-type ovsdbEventHandler interface {
-	AddLocalEndpoint(endpoint datapath.Endpoint)
-	DeleteLocalEndpoint(endpoint datapath.Endpoint)
-	UpdateLocalEndpoint(newEndpoint datapath.Endpoint, oldEndpoint datapath.Endpoint)
-}
-
-type OvsdbEventHandlerFuncs struct {
-	LocalEndpointAddFunc    func(endpoint datapath.Endpoint)
-	LocalEndpointDeleteFunc func(endpoint datapath.Endpoint)
-	LocalEndpointUpdateFunc func(newEndpoint datapath.Endpoint, oldEndpoint datapath.Endpoint)
-}
-
-func (handler OvsdbEventHandlerFuncs) AddLocalEndpoint(endpoint datapath.Endpoint) {
-	if handler.LocalEndpointAddFunc != nil {
-		handler.LocalEndpointAddFunc(endpoint)
-	}
-}
-
-func (handler OvsdbEventHandlerFuncs) DeleteLocalEndpoint(endpoint datapath.Endpoint) {
-	if handler.LocalEndpointDeleteFunc != nil {
-		handler.LocalEndpointDeleteFunc(endpoint)
-	}
-}
-
-func (handler OvsdbEventHandlerFuncs) UpdateLocalEndpoint(newEndpoint datapath.Endpoint, oldEndpoint datapath.Endpoint) {
-	if handler.LocalEndpointUpdateFunc != nil {
-		handler.LocalEndpointUpdateFunc(newEndpoint, oldEndpoint)
-	}
-}
-
-func (monitor *AgentMonitor) RegisterOvsdbEventHandler(ovsdbEventHandler ovsdbEventHandler) {
-	if ovsdbEventHandler == nil {
-		klog.Fatalf("Failed to register ovsdbEventHandler: register nil ovsdbEventHandler not allow")
-	}
-	if monitor.ovsdbEventHandler != nil {
-		klog.Fatalf("Failed to register ovsdbEventHandler: monitor ovsdbEventHandler already register")
-	}
-
-	monitor.ovsdbEventHandler = ovsdbEventHandler
-}
-
-// agentMonitor monitor agent state, update agentinfo to apiserver.
+// AgentMonitor monitor agent state, update agentinfo to apiserver.
 type AgentMonitor struct {
 	// k8sClient used to create/read/update agentinfo
 	k8sClient client.Client
-	// ovsClient used to monitor ovsdb table port/bridge/interface
-	ovsClient *ovsdb.OvsdbClient
+	// ovsdbMonitor used to access ovsdb cache
+	ovsdbMonitor *OVSDBMonitor
 
 	// agentName is the name and uuid of this agent
-	agentName string
-
-	// cacheLock is a read/write lock for accessing the cache
-	cacheLock                  sync.RWMutex
-	ovsdbCache                 map[string]map[string]ovsdb.Row
+	agentName                  string
 	ipCacheLock                sync.RWMutex
 	ipCache                    map[string]map[types.IPAddress]metav1.Time
 	ofPortIPAddressMonitorChan chan map[string]net.IP
-
-	ovsdbEventHandler                  ovsdbEventHandler
-	localEndpointHardwareAddrCacheLock sync.RWMutex
-	localEndpointHardwareAddrCache     map[string]uint32
 
 	// syncQueue used to notify agentMonitor synchronize AgentInfo
 	syncQueue workqueue.RateLimitingInterface
 }
 
 // NewAgentMonitor return a new agentMonitor with kubernetes client and ipMonitor.
-func NewAgentMonitor(client client.Client, ofPortIPAddressMonitorChan chan map[string]net.IP) (*AgentMonitor, error) {
-	monitor := &AgentMonitor{
-		k8sClient:                          client,
-		cacheLock:                          sync.RWMutex{},
-		ipCacheLock:                        sync.RWMutex{},
-		ovsdbCache:                         make(map[string]map[string]ovsdb.Row),
-		ipCache:                            make(map[string]map[types.IPAddress]metav1.Time),
-		ofPortIPAddressMonitorChan:         ofPortIPAddressMonitorChan,
-		localEndpointHardwareAddrCache:     make(map[string]uint32),
-		localEndpointHardwareAddrCacheLock: sync.RWMutex{},
-		syncQueue:                          workqueue.NewRateLimitingQueue(workqueue.DefaultItemBasedRateLimiter()),
+func NewAgentMonitor(client client.Client, ovsdbMonitor *OVSDBMonitor, ofPortIPAddressMonitorChan chan map[string]net.IP) *AgentMonitor {
+	return &AgentMonitor{
+		k8sClient:                  client,
+		agentName:                  utils.CurrentAgentName(),
+		ipCacheLock:                sync.RWMutex{},
+		ipCache:                    make(map[string]map[types.IPAddress]metav1.Time),
+		ofPortIPAddressMonitorChan: ofPortIPAddressMonitorChan,
+		ovsdbMonitor:               ovsdbMonitor,
+		syncQueue:                  ovsdbMonitor.GetSyncQueue(),
 	}
-
-	var err error
-
-	monitor.agentName = utils.CurrentAgentName()
-
-	monitor.ovsClient, err = ovsdb.ConnectUnix(ovsdb.DEFAULT_SOCK)
-	if err != nil {
-		klog.Errorf("failed to connection to ovsdb: %s", err.Error())
-		return nil, err
-	}
-
-	return monitor, nil
 }
 
 func (monitor *AgentMonitor) Run(stopChan <-chan struct{}) {
 	defer monitor.syncQueue.ShutDown()
-	defer monitor.ovsClient.Disconnect()
 
 	klog.Infof("start agent %s monitor", monitor.Name())
 	defer klog.Infof("shutting down agent %s monitor", monitor.Name())
 
-	var err error
-	err = monitor.startOvsdbMonitor()
-	if err != nil {
-		klog.Fatalf("unable start ovsdb monitor: %s", err)
-	}
-	go monitor.HandleOfPortIPAddressUpdate(monitor.ofPortIPAddressMonitorChan, stopChan)
-
+	go monitor.handleOfPortIPAddressUpdate(monitor.ofPortIPAddressMonitorChan, stopChan)
 	go wait.Until(monitor.syncAgentInfoWorker, 0, stopChan)
 	go monitor.periodicallySyncAgentInfo(AgentInfoSyncInterval, stopChan)
 	<-stopChan
 }
 
-func (monitor *AgentMonitor) HandleOfPortIPAddressUpdate(ofPortIPAddressMonitorChan <-chan map[string]net.IP, stopChan <-chan struct{}) {
+func (monitor *AgentMonitor) handleOfPortIPAddressUpdate(ofPortIPAddressMonitorChan <-chan map[string]net.IP, stopChan <-chan struct{}) {
 	for {
 		select {
 		case localEndpointInfo := <-ofPortIPAddressMonitorChan:
@@ -194,72 +122,6 @@ func (monitor *AgentMonitor) periodicallySyncAgentInfo(cycle int, stopChan <-cha
 		case <-stopChan:
 			return
 		}
-	}
-}
-
-func (monitor *AgentMonitor) startOvsdbMonitor() error {
-	klog.Infof("start monitor ovsdb %s", "Open_vSwitch")
-	monitor.ovsClient.Register(ovsUpdateHandlerFunc(monitor.handleOvsUpdates))
-
-	selectAll := ovsdb.MonitorSelect{
-		Initial: true,
-		Insert:  true,
-		Delete:  true,
-		Modify:  true,
-	}
-	requests := map[string]ovsdb.MonitorRequest{
-		"Port":         {Select: selectAll, Columns: []string{"name", "interfaces", "external_ids", "bond_mode", "vlan_mode", "tag", "trunks"}},
-		"Interface":    {Select: selectAll, Columns: []string{"name", "mac_in_use", "ofport", "type", "external_ids", "error"}},
-		"Bridge":       {Select: selectAll, Columns: []string{"name", "ports"}},
-		"Open_vSwitch": {Select: selectAll, Columns: []string{"ovs_version"}},
-	}
-
-	initial, err := monitor.ovsClient.Monitor("Open_vSwitch", nil, requests)
-	if err != nil {
-		return fmt.Errorf("monitor ovsdb %s: %s", "Open_vSwitch", err)
-	}
-	monitor.handleOvsUpdates(*initial)
-
-	return nil
-}
-
-// Endpoint implement in everoute datapath module pr
-func (monitor *AgentMonitor) interfaceToEndpoint(ofport uint32, interfaceName, macAddrStr string) *datapath.Endpoint {
-	// NOTE should use interface uuid to caculate endpoint info
-	var bridgeName string
-	var portUUID string
-	var vlanID uint16
-
-	monitor.cacheLock.RLock()
-	defer monitor.cacheLock.RUnlock()
-	for uuid, port := range monitor.ovsdbCache["Port"] {
-		if port.Fields["name"].(string) == interfaceName {
-			portUUID = uuid
-			tag, ok := port.Fields["tag"].(float64)
-			if !ok {
-				break
-			}
-			vlanID = uint16(tag)
-			break
-		}
-	}
-
-	for _, bridge := range monitor.ovsdbCache["Bridge"] {
-		portUUIDs := listUUID(bridge.Fields["ports"])
-		for _, uuid := range portUUIDs {
-			if uuid.GoUuid == portUUID {
-				bridgeName = bridge.Fields["name"].(string)
-				break
-			}
-		}
-	}
-
-	return &datapath.Endpoint{
-		InterfaceName: interfaceName,
-		MacAddrStr:    macAddrStr,
-		PortNo:        ofport,
-		BridgeName:    bridgeName,
-		VlanID:        vlanID,
 	}
 }
 
@@ -338,9 +200,6 @@ func (monitor *AgentMonitor) mergeAgentInfo(localAgentInfo, cpAgentInfo *agentv1
 }
 
 func (monitor *AgentMonitor) getAgentInfo() (*agentv1alpha1.AgentInfo, error) {
-	monitor.cacheLock.RLock()
-	defer monitor.cacheLock.RUnlock()
-
 	agentInfo := &agentv1alpha1.AgentInfo{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      monitor.Name(),
@@ -348,22 +207,28 @@ func (monitor *AgentMonitor) getAgentInfo() (*agentv1alpha1.AgentInfo, error) {
 		},
 	}
 
-	ovsVersion, err := monitor.fetchOvsVersionLocked()
-	if err == nil {
-		agentInfo.OVSInfo.Version = ovsVersion
-	}
-
 	hostname, err := os.Hostname()
 	if err == nil {
 		agentInfo.Hostname = hostname
 	}
 
-	for uuid := range monitor.ovsdbCache["Bridge"] {
-		bridge, err := monitor.fetchBridgeLocked(ovsdb.UUID{GoUuid: uuid})
-		if err != nil {
-			return nil, fmt.Errorf("unable fetch bridge %s: %s", uuid, err)
+	err = monitor.ovsdbMonitor.LockedAccessCache(func(ovsdbCache OVSDBCache) error {
+		ovsVersion, err := monitor.fetchOvsVersionLocked(ovsdbCache)
+		if err == nil {
+			agentInfo.OVSInfo.Version = ovsVersion
 		}
-		agentInfo.OVSInfo.Bridges = append(agentInfo.OVSInfo.Bridges, *bridge)
+
+		for uuid := range ovsdbCache["Bridge"] {
+			bridge, err := monitor.fetchBridgeLocked(ovsdbCache, ovsdb.UUID{GoUuid: uuid})
+			if err != nil {
+				return fmt.Errorf("unable fetch bridge %s: %s", uuid, err)
+			}
+			agentInfo.OVSInfo.Bridges = append(agentInfo.OVSInfo.Bridges, *bridge)
+		}
+		return nil
+	})
+	if err != nil {
+		return nil, err
 	}
 
 	agentHealthCondition := agentv1alpha1.AgentCondition{
@@ -376,291 +241,12 @@ func (monitor *AgentMonitor) getAgentInfo() (*agentv1alpha1.AgentInfo, error) {
 	return agentInfo, nil
 }
 
-func (monitor *AgentMonitor) filterEndpoint(rowupdate ovsdb.RowUpdate) (*datapath.Endpoint, *datapath.Endpoint) {
-	newExternalIds := rowupdate.New.Fields["external_ids"].(ovsdb.OvsMap).GoMap
-	_, ok := newExternalIds[LocalEndpointIdentity]
-	if !ok {
-		if rowupdate.New.Fields["mac_in_use"] == "" {
-			return nil, nil
-		}
-	}
-	return monitor.filterLocalEndpoint(rowupdate)
-}
-
-func (monitor *AgentMonitor) filterLocalEndpoint(rowupdate ovsdb.RowUpdate) (*datapath.Endpoint, *datapath.Endpoint) {
-	empty := ovsdb.Row{}
-	if reflect.DeepEqual(rowupdate.Old, empty) {
-		return monitor.filterEndpointAdded(rowupdate), nil
-	}
-
-	if rowupdate.Old.Fields["external_ids"] == nil {
-		return monitor.filterEndpointAdded(rowupdate), nil
-	}
-
-	return monitor.filterEndpointUpdated(rowupdate)
-}
-
-func (monitor *AgentMonitor) filterEndpointUpdated(rowupdate ovsdb.RowUpdate) (*datapath.Endpoint, *datapath.Endpoint) {
-	var macStr string
-	newExternalIds := rowupdate.New.Fields["external_ids"].(ovsdb.OvsMap).GoMap
-	_, ok := newExternalIds[LocalEndpointIdentity]
-	if !ok {
-		macStr, _ = rowupdate.New.Fields["mac_in_use"].(string)
-	} else {
-		macStr, _ = newExternalIds[LocalEndpointIdentity].(string)
-	}
-	newOfPort, ok := rowupdate.New.Fields["ofport"].(float64)
-	if !ok {
-		return nil, nil
-	}
-	oldOfPort, ok := rowupdate.Old.Fields["ofport"].(float64)
-	if !ok {
-		return nil, nil
-	}
-
-	if newOfPort <= 0 || oldOfPort <= 0 {
-		return nil, nil
-	}
-	macAddr, err := net.ParseMAC(macStr)
-	if err != nil {
-		klog.Errorf("Parsing endpoint macAddr error: %v", macAddr)
-		return nil, nil
-	}
-	// interface udpate is triggerd by some other fileds except that ofport field
-	if newOfPort == oldOfPort {
-		return nil, nil
-	}
-	monitor.localEndpointHardwareAddrCacheLock.Lock()
-	defer monitor.localEndpointHardwareAddrCacheLock.Unlock()
-	monitor.localEndpointHardwareAddrCache[macStr] = uint32(newOfPort)
-
-	if err := monitor.waitForPortPresent(macStr, 1*time.Second); err != nil {
-		klog.Errorf("Failed to update local endpoint, wait ovs port of newEndpoint present timeout, error: %v", err)
-		return nil, nil
-	}
-
-	newEndpoint := monitor.interfaceToEndpoint(uint32(newOfPort), rowupdate.New.Fields["name"].(string), macAddr.String())
-	oldEndpoint := monitor.interfaceToEndpoint(uint32(oldOfPort), rowupdate.Old.Fields["name"].(string), macAddr.String())
-	return newEndpoint, oldEndpoint
-}
-
-func (monitor *AgentMonitor) filterEndpointAdded(rowupdate ovsdb.RowUpdate) *datapath.Endpoint {
-	var macStr string
-	newExternalIds := rowupdate.New.Fields["external_ids"].(ovsdb.OvsMap).GoMap
-	_, ok := newExternalIds[LocalEndpointIdentity]
-	if !ok {
-		macStr, _ = rowupdate.New.Fields["mac_in_use"].(string)
-	} else {
-		macStr, _ = newExternalIds[LocalEndpointIdentity].(string)
-	}
-	monitor.localEndpointHardwareAddrCacheLock.Lock()
-	defer monitor.localEndpointHardwareAddrCacheLock.Unlock()
-
-	if _, ok := monitor.localEndpointHardwareAddrCache[macStr]; ok {
-		return nil
-	}
-	ofPort, ok := rowupdate.New.Fields["ofport"].(float64)
-	if !ok {
-		return nil
-	}
-
-	if ofPort <= 0 {
-		// Parsing added ofport error: invalid local endpoint ofPort. In OfPort initializing status
-		return nil
-	}
-	ofport := uint32(ofPort)
-
-	macAddr, err := net.ParseMAC(macStr)
-	if err != nil {
-		klog.Errorf("Parsing endpoint macAddr error: %v", macAddr)
-		return nil
-	}
-	monitor.localEndpointHardwareAddrCache[macStr] = uint32(ofPort)
-
-	if err := monitor.waitForPortPresent(rowupdate.New.Fields["name"].(string), 1*time.Second); err != nil {
-		klog.Errorf("Failed to add local endpoint, wait ovs port present timeout, error: %v", err)
-		return nil
-	}
-
-	return monitor.interfaceToEndpoint(ofport, rowupdate.New.Fields["name"].(string), macStr)
-}
-
-func (monitor *AgentMonitor) waitForPortPresent(interfaceName string, timeout time.Duration) error {
-	return wait.PollImmediate(10*time.Millisecond, timeout, func() (do bool, err error) {
-		if !monitor.isPortExists(interfaceName) {
-			return false, nil
-		}
-
-		return true, nil
-	})
-}
-
-func (monitor *AgentMonitor) isPortExists(interfaceName string) bool {
-	monitor.cacheLock.RLock()
-	defer monitor.cacheLock.RUnlock()
-	for _, port := range monitor.ovsdbCache["Port"] {
-		if port.Fields["name"].(string) == interfaceName {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (monitor *AgentMonitor) filterEndpointDeleted(rowupdate ovsdb.RowUpdate) *datapath.Endpoint {
-	var ofport uint32
-	var macStr string
-	oldExternalIds := rowupdate.Old.Fields["external_ids"].(ovsdb.OvsMap).GoMap
-	_, ok := oldExternalIds[LocalEndpointIdentity]
-	if !ok {
-		macStr, _ = rowupdate.Old.Fields["mac_in_use"].(string)
-	} else {
-		macStr, _ = oldExternalIds[LocalEndpointIdentity].(string)
-	}
-
-	monitor.localEndpointHardwareAddrCacheLock.Lock()
-	defer monitor.localEndpointHardwareAddrCacheLock.Unlock()
-
-	if _, ok := monitor.localEndpointHardwareAddrCache[macStr]; !ok {
-		return nil
-	}
-
-	ofPort, ok := rowupdate.Old.Fields["ofport"].(float64)
-	if !ok {
-		return nil
-	}
-	if ofPort <= 0 {
-		ofport = monitor.localEndpointHardwareAddrCache[macStr]
-	} else {
-		ofport = uint32(ofPort)
-	}
-
-	delete(monitor.localEndpointHardwareAddrCache, macStr)
-
-	return monitor.interfaceToEndpoint(ofport, rowupdate.Old.Fields["name"].(string), macStr)
-}
-
-func (monitor *AgentMonitor) filterPortVlanTagUpdate(rowupdate ovsdb.RowUpdate) (*datapath.Endpoint, *datapath.Endpoint) {
-	monitor.cacheLock.Lock()
-	ok, localEndpointIface := monitor.isLocalEndpointPort(rowupdate.New.Fields["name"].(string))
-	if !ok {
-		monitor.cacheLock.Unlock()
-		return nil, nil
-	}
-	monitor.cacheLock.Unlock()
-
-	var oldTag, newTag float64
-	var ofport uint32
-	if rowupdate.New.Fields["tag"] == nil {
-		return nil, nil
-	}
-	if rowupdate.Old.Fields["tag"] == nil {
-		return nil, nil
-	}
-	newTag, _ = rowupdate.New.Fields["tag"].(float64)
-	oldTag, _ = rowupdate.Old.Fields["tag"].(float64)
-	if newTag == oldTag {
-		return nil, nil
-	}
-
-	portName := localEndpointIface.Fields["name"].(string)
-	externalIDs := localEndpointIface.Fields["external_ids"].(ovsdb.OvsMap).GoMap
-	ofPort, ok := localEndpointIface.Fields["ofport"].(float64)
-	if !ok {
-		return nil, nil
-	}
-	if ofPort <= 0 {
-		monitor.localEndpointHardwareAddrCacheLock.Lock()
-		ofport = monitor.localEndpointHardwareAddrCache[externalIDs[LocalEndpointIdentity].(string)]
-		monitor.localEndpointHardwareAddrCacheLock.Unlock()
-	} else {
-		ofport = uint32(ofPort)
-	}
-
-	newEndpoint := monitor.interfaceToEndpoint(ofport, portName, externalIDs[LocalEndpointIdentity].(string))
-	oldEndpoint := &datapath.Endpoint{
-		InterfaceName: newEndpoint.InterfaceName,
-		MacAddrStr:    newEndpoint.MacAddrStr,
-		PortNo:        newEndpoint.PortNo,
-		BridgeName:    newEndpoint.BridgeName,
-		VlanID:        uint16(oldTag),
-	}
-
-	return newEndpoint, oldEndpoint
-}
-
-func (monitor *AgentMonitor) isLocalEndpointPort(portName string) (bool, ovsdb.Row) {
-	for i, iface := range monitor.ovsdbCache["Interface"] {
-		if iface.Fields["name"].(string) == portName {
-			if _, ok := iface.Fields["external_ids"].(ovsdb.OvsMap).GoMap[LocalEndpointIdentity]; ok {
-				return true, monitor.ovsdbCache["Interface"][i]
-			}
-		}
-	}
-	return false, ovsdb.Row{}
-}
-
-func (monitor *AgentMonitor) processPortUpdate(rowupdate ovsdb.RowUpdate) {
-	newEndpoint, oldEndpoint := monitor.filterPortVlanTagUpdate(rowupdate)
-	if newEndpoint != nil && oldEndpoint != nil {
-		go monitor.ovsdbEventHandler.UpdateLocalEndpoint(*newEndpoint, *oldEndpoint)
-	}
-}
-
-func (monitor *AgentMonitor) processEndpointUpdate(rowupdate ovsdb.RowUpdate) {
-	newEndpoint, oldEndpoint := monitor.filterEndpoint(rowupdate)
-	if newEndpoint != nil && oldEndpoint != nil {
-		go monitor.ovsdbEventHandler.UpdateLocalEndpoint(*newEndpoint, *oldEndpoint)
-	} else if newEndpoint != nil {
-		go monitor.ovsdbEventHandler.AddLocalEndpoint(*newEndpoint)
-	}
-}
-
-func (monitor *AgentMonitor) processEndpointDel(rowupdate ovsdb.RowUpdate) {
-	deletedEndpoints := monitor.filterEndpointDeleted(rowupdate)
-	if deletedEndpoints != nil {
-		go monitor.ovsdbEventHandler.DeleteLocalEndpoint(*deletedEndpoints)
-	}
-}
-
-func (monitor *AgentMonitor) handleOvsUpdates(updates ovsdb.TableUpdates) {
-	monitor.cacheLock.Lock()
-	defer monitor.cacheLock.Unlock()
-
-	for table, tableUpdate := range updates.Updates {
-		if _, ok := monitor.ovsdbCache[table]; !ok {
-			monitor.ovsdbCache[table] = make(map[string]ovsdb.Row)
-		}
-		for uuid, row := range tableUpdate.Rows {
-			empty := ovsdb.Row{}
-			if !reflect.DeepEqual(row.New, empty) {
-				if table == "Interface" {
-					go monitor.processEndpointUpdate(row)
-				}
-				if table == "Port" {
-					go monitor.processPortUpdate(row)
-				}
-
-				monitor.ovsdbCache[table][uuid] = row.New
-			} else {
-				if table == "Interface" {
-					go monitor.processEndpointDel(row)
-				}
-
-				delete(monitor.ovsdbCache[table], uuid)
-			}
-		}
-	}
-
-	monitor.syncQueue.Add(monitor.Name())
-}
-
 func (monitor *AgentMonitor) Name() string {
 	return monitor.agentName
 }
 
-func (monitor *AgentMonitor) fetchOvsVersionLocked() (string, error) {
-	tableOvs := monitor.ovsdbCache["Open_vSwitch"]
+func (monitor *AgentMonitor) fetchOvsVersionLocked(ovsdbCache OVSDBCache) (string, error) {
+	tableOvs := ovsdbCache["Open_vSwitch"]
 	if len(tableOvs) == 0 {
 		return "", fmt.Errorf("couldn't find table %s, agentMonitor may haven't start", "Open_vSwitch")
 	}
@@ -672,8 +258,8 @@ func (monitor *AgentMonitor) fetchOvsVersionLocked() (string, error) {
 	return "", nil
 }
 
-func (monitor *AgentMonitor) fetchPortLocked(uuid ovsdb.UUID, bridgeName string) (*agentv1alpha1.OVSPort, error) {
-	ovsPort, ok := monitor.ovsdbCache["Port"][uuid.GoUuid]
+func (monitor *AgentMonitor) fetchPortLocked(ovsdbCache OVSDBCache, uuid ovsdb.UUID, bridgeName string) (*agentv1alpha1.OVSPort, error) {
+	ovsPort, ok := ovsdbCache["Port"][uuid.GoUuid]
 	if !ok {
 		return nil, fmt.Errorf("ovs port %s not found in cache", uuid)
 	}
@@ -707,7 +293,7 @@ func (monitor *AgentMonitor) fetchPortLocked(uuid ovsdb.UUID, bridgeName string)
 	}
 
 	for _, uuid := range listUUID(ovsPort.Fields["interfaces"]) {
-		iface := monitor.fetchInterfaceLocked(uuid, bridgeName)
+		iface := monitor.fetchInterfaceLocked(ovsdbCache, uuid, bridgeName)
 		if iface != nil {
 			port.Interfaces = append(port.Interfaces, *iface)
 		}
@@ -716,8 +302,8 @@ func (monitor *AgentMonitor) fetchPortLocked(uuid ovsdb.UUID, bridgeName string)
 	return port, nil
 }
 
-func (monitor *AgentMonitor) fetchInterfaceLocked(uuid ovsdb.UUID, bridgeName string) *agentv1alpha1.OVSInterface {
-	ovsIface, ok := monitor.ovsdbCache["Interface"][uuid.GoUuid]
+func (monitor *AgentMonitor) fetchInterfaceLocked(ovsdbCache OVSDBCache, uuid ovsdb.UUID, bridgeName string) *agentv1alpha1.OVSInterface {
+	ovsIface, ok := ovsdbCache["Interface"][uuid.GoUuid]
 	if !ok {
 		klog.V(4).Infof("could not find interface %+v in cache", ovsIface)
 		return nil
@@ -756,8 +342,8 @@ func (monitor *AgentMonitor) fetchInterfaceLocked(uuid ovsdb.UUID, bridgeName st
 	return &iface
 }
 
-func (monitor *AgentMonitor) fetchBridgeLocked(uuid ovsdb.UUID) (*agentv1alpha1.OVSBridge, error) {
-	ovsBri, ok := monitor.ovsdbCache["Bridge"][uuid.GoUuid]
+func (monitor *AgentMonitor) fetchBridgeLocked(ovsdbCache OVSDBCache, uuid ovsdb.UUID) (*agentv1alpha1.OVSBridge, error) {
+	ovsBri, ok := ovsdbCache["Bridge"][uuid.GoUuid]
 	if !ok {
 		return nil, fmt.Errorf("ovs bridge %s not found in cache", uuid)
 	}
@@ -767,7 +353,7 @@ func (monitor *AgentMonitor) fetchBridgeLocked(uuid ovsdb.UUID) (*agentv1alpha1.
 	}
 
 	for _, uuid := range listUUID(ovsBri.Fields["ports"]) {
-		port, err := monitor.fetchPortLocked(uuid, bridge.Name)
+		port, err := monitor.fetchPortLocked(ovsdbCache, uuid, bridge.Name)
 		if err != nil {
 			return nil, err
 		}
